# apps/docs-platform/public/openapi.yaml
openapi: 3.1.0
info:
  title: VeriWorkly API
  version: 1.0.0
  description: |
    The official API reference for VeriWorkly.

    ### Authentication Methods

    VeriWorkly supports two authentication methods:

    #### 1. API Key Authentication (recommended for scripts and integrations)
    For programmatic access, send your key either way — both are accepted:

    - `X-API-Key: vw_...`
    - `Authorization: Bearer vw_...`

    A key is **required** for any request that does not originate from a whitelisted VeriWorkly
    origin, including requests made from these documentation pages.

    - **Format**: `vw_` followed by 64 hexadecimal characters.
    - **Rate limit**: 20 requests per 15 minutes per key by default, configurable per key at
      creation. Responses carry `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and
      `X-RateLimit-Reset`.
    - **Scopes**: keys are scope-gated. A key missing a route's required scope gets `403`. Session
      callers are never scope-gated.

    #### 2. Session Authentication (dashboard only)
    Used internally by the VeriWorkly dashboard and resume builder, via secure HTTP-only cookies
    issued by Better Auth.

    - **Cookies**:
      - Local: `veriworkly-auth.session_token`
      - Production: `__Secure-veriworkly-auth.session_token`
    - **Note**: sessions are only honoured for whitelisted origins. Send `credentials: 'include'`
      (fetch) or `withCredentials: true` (axios). Requests from these documentation pages **must**
...

...

...
servers:
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      operationId: getCurrentUser
      security:
        - ApiKeyAuth: []
        - LocalSession: []
        - ProdSession: []
      summary: Get Current User
      description: |
        Retrieves account metadata for the currently authenticated user. Requires the `user:read`
        scope when called with an API key.

        This returns identity metadata only — resume, cover letter, and portfolio content is never
        exposed here. Results are cached for 30 minutes and refreshed on write.
      responses:
        "200":
          description: User information fetched successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                success: true
                message: User information fetched successfully
                data:
                  id: cmnu7awv4000063lk18zpon2m
                  name: VeriWorkly User
                  username: veriworkly-user
                  email: info@veriworkly.com
                  emailVerified: true
                  autoSyncEnabled: false
                  createdAt: 2026-04-11T10:37:35.584Z
                  updatedAt: 2026-04-25T19:35:44.409Z
                  _count:
                    apiKeys: 2
                    shareLinks: 4
                    resumes: 4
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                statusCode: 404
                message: User not found
  /api/v1/users/me/name:
    put:
      tags:
        - Users
      operationId: updateUserName
      security:
        - ApiKeyAuth: []
        - LocalSession: []
        - ProdSession: []
      summary: Update User Name
      description: |
        Updates the display name of the authenticated user. Requires the `user:write` scope when
        called with an API key.

        The email address is read-only through this API. Leading and trailing whitespace is trimmed
        before validation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Name cannot be null or empty when updating.
            example:
              name: VeriWorkly
      responses:
        "200":
          description: User name updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                success: true
                message: User name updated successfully
                data:
                  id: cmnu7awv4000063lk18zpon2m
                  name: VeriWorkly User New Name
                  username: veriworkly-user
                  email: info@veriworkly.com
                  emailVerified: true
                  autoSyncEnabled: false
                  createdAt: 2026-04-11T10:37:35.584Z
                  updatedAt: 2026-04-26T02:00:00.000Z
                  _count:
                    apiKeys: 2
                    shareLinks: 4
                    resumes: 4
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                emptyName:
                  summary: Name is empty
                  value:
                    success: false
                    statusCode: 400
                    message: Validation failed
                    details:
                      - path: name
                        message: Name cannot be empty
                nameTooLong:
                  summary: Name too long
                  value:
                    success: false
                    statusCode: 400
                    message: Validation failed
                    details:
                      - path: name
                        message: Name is too long
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v1/users/me/sync:
    put:
      tags:
        - Users
      operationId: updateAutoSync
      security:
        - ApiKeyAuth: []
        - LocalSession: []
        - ProdSession: []
      summary: Update Auto-Sync Preference
      description: |
        Turns the account-level background document sync on or off. Requires the `user:write` scope
        when called with an API key.

        This is the account-wide switch. It does not override a document's own `keepLocalOnly` flag —
        a document opted out individually stays local even with auto-sync on. Turning auto-sync off
        does not delete anything already synced; it only stops new local edits from being pushed.

        Defaults to `true` on a new account. A successful change invalidates the caller's cached
        session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - enabled
              properties:
                enabled:
                  type: boolean
            example:
              enabled: true
      responses:
        "200":
          description: Auto sync preference updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                success: true
                message: Auto sync preference updated successfully
                data:
                  id: cmnu7awv4000063lk18zpon2m
                  name: VeriWorkly User
                  username: veriworkly-user
                  email: info@veriworkly.com
                  emailVerified: true
                  autoSyncEnabled: true
                  createdAt: 2026-04-11T10:37:35.584Z
                  updatedAt: 2026-04-26T02:00:00.000Z
                  _count:
                    apiKeys: 2
                    shareLinks: 4
                    resumes: 4
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                statusCode: 400
                message: Validation failed
                details:
                  - path: enabled
                    message: Expected boolean, received string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/users/{username}/availability:
    get:
      tags:
        - Users
      operationId: checkUsernameAvailability
      security:
        - ApiKeyAuth: []
        - LocalSession: []
        - ProdSession: []
      summary: Check Username Availability
      description: |
        Reports whether a username can be claimed. Intended to drive live feedback in a username
        picker before `PUT /users/me/username` is called.

        This route carries **no API-key scope gate**, unlike the rest of the Users module — but it is
        still not anonymous, so it needs an API key or a whitelisted first-party origin.

        The path segment is normalised the same way the update endpoint normalises it, and
        `normalizedUsername` echoes back the value that would actually be stored. A `200` with
        `available: false` is the normal negative answer — an unavailable name is not an error.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The candidate username. Normalised server-side before the lookup.
      responses:
        "200":
          description: Username availability fetched successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailabilityResponse"
              examples:
                available:
                  summary: Free to claim
                  value:
                    success: true
                    message: Username availability fetched successfully
                    data:
                      available: true
                      normalizedUsername: veriworkly-user
                taken:
                  summary: Already claimed
                  value:
                    success: true
                    message: Username availability fetched successfully
                    data:
                      available: false
                      normalizedUsername: veriworkly-user
                      reason: taken
                invalid:
                  summary: Fails the username rules
                  value:
                    success: true
                    message: Username availability fetched successfully
                    data:
                      available: false
                      normalizedUsername: a
                      reason: too_short
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v1/github/stats:
    get:
      tags:
        - GitHub
      operationId: getGitHubStats
      summary: Get GitHub Stats
      security:
        - ApiKeyAuth: []
        - LocalSession: []
        - ProdSession: []
      description: |
        Fetch aggregated issue and pull-request statistics for the configured repository, as of the
        last sync. Served from the last synced snapshot, not live from GitHub — `syncedAt` tells you
        how fresh it is. Cached for 12 hours.

        This endpoint has no user session of its own: call it with an API key holding the
        `github:read` scope, or from a whitelisted first-party origin. A request with neither is
        rejected with `401`.
      responses:
        "200":
          description: GitHub stats fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubStatsResponse"
              examples:
                synced:
                  summary: Stats available
                  value:
                    success: true
                    message: GitHub stats fetched successfully
                    data:
                      projectName: owner/repo
                      stats:
                        total: 120
                        issues: 70
                        pullRequests: 50
                        todo: 40
                        inProgress: 30
                        done: 50
                        completionRate: "41.67"
                      syncedAt: 2026-04-26T10:00:00.000Z
                neverSynced:
                  summary: No sync has run yet
                  value:
                    success: true
                    message: GitHub stats fetched successfully
                    data: null
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Your personal API key — `vw_` followed by 64 hex characters. The same key is also accepted
        as `Authorization: Bearer vw_...`.
    LocalSession:
      type: apiKey
      in: cookie
      name: veriworkly-auth.session_token
      description: Internal session cookie for local development.
    ProdSession:
      type: apiKey
      in: cookie
      name: __Secure-veriworkly-auth.session_token
      description: Internal session cookie for production.
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - statusCode
      properties:
        success:
          type: boolean
        message:
          type: string
        statusCode:
          type: integer
        details:
          type: array
          nullable: true
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
...
    UserObject:
      type: object
      required:
        - id
        - email
        - emailVerified
        - autoSyncEnabled
        - createdAt
        - updatedAt
        - _count
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
          description: |
            Null until the user claims one. Immutable once set, and required before any share link can
            be created.
        email:
          type: string
        emailVerified:
          type: boolean
        autoSyncEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        _count:
          type: object
          description: Relational counts for the user's own records.
          required:
            - apiKeys
            - shareLinks
            - resumes
          properties:
            apiKeys:
              type: integer
            shareLinks:
              type: integer
            resumes:
              type: integer
              description: |
                Counts every Document the user owns — every `type`, and soft-deleted ones included.
                It is not a count of RESUME-type documents. Use `GET /documents` for an accurate,
                filterable list.
    UserResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserObject"
    UsernameAvailability:
      type: object
      required:
        - available
        - normalizedUsername
      properties:
        available:
          type: boolean
        normalizedUsername:
          type: string
          description: The value that would actually be stored, after server-side normalisation.
        reason:
          type: string
          enum:
            - too_short
            - too_long
            - invalid_characters
            - reserved
            - taken
          description: Present only when `available` is `false`.
    UsernameAvailabilityResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UsernameAvailability"
...
    GitHubStats:
      type: object
      properties:
        total:
          type: integer
          description: Issues plus pull requests.
        issues:
          type: integer
          description: Issue-only count.
        pullRequests:
          type: integer
        todo:
          type: integer
        inProgress:
          type: integer
        done:
          type: integer
        completionRate:
          type: string
          description: Percentage of `total` that is done, formatted to 2 decimal places. `"0.00"` when there is nothing synced.
          example: "41.67"
    GitHubStatsData:
      type: object
      properties:
        projectName:
          type: string
        stats:
          $ref: "#/components/schemas/GitHubStats"
        syncedAt:
          type: string
          format: date-time
    GitHubStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Null until the first GitHub sync has run for the configured repository.
          oneOf:
            - $ref: "#/components/schemas/GitHubStatsData"
            - type: "null"
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized - Authentication failed or session expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            statusCode: 401
            message: Authentication required
...